requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.10
langchain-community==0.0.29
faiss-cpu==1.7.4
sentence-transformers==2.6.1
pydub==0.25.1
websockets==12.0
openai-whisper==20231117
torch==2.2.1
torchaudio==2.2.1
ollama==0.1.4
gTTS==2.5.0
python-multipart==0.0.6
psutil==5.9.5
soundfile==0.12.1
numpy==1.26.4
librosa==0.10.1